pass all tests in candidate_backing_test file
