+++ dot/parachain/backing/candidate_backing_test.go
@@ -112,6 +112,127 @@ func secondedSignedFullStatementWithPVD(
 	}
 }
 
+func TestImportStatement(t *testing.T) {
+	statementVDTValid := parachaintypes.NewStatementVDT()
+	err := statementVDTValid.SetValue(parachaintypes.Valid{})
+	require.NoError(t, err)
+
+	dummyCCR := getDummyCommittedCandidateReceipt(t)
+	seconded := parachaintypes.Seconded(dummyCCR)
+
+	candidateHash, err := parachaintypes.GetCandidateHash(dummyCCR)
+	require.NoError(t, err)
+
+	statementVDTSeconded := parachaintypes.NewStatementVDT()
+	err = statementVDTSeconded.SetValue(seconded)
+	require.NoError(t, err)
+
+	testCases := []struct {
+		description            string
+		rpState                func() perRelayParentState
+		perCandidate           map[parachaintypes.CandidateHash]*perCandidateState
+		signedStatementWithPVD parachaintypes.SignedFullStatementWithPVD
+		mockOverseer           func(*testing.T, chan any)
+	}{
+		{
+			description: "statement_is_not_seconded",
+			rpState: func() perRelayParentState {
+				ctrl := gomock.NewController(t)
+				mockTable := NewMockTable(ctrl)
+
+				mockTable.EXPECT().importStatement(
+					gomock.AssignableToTypeOf(new(tableContext)),
+					gomock.AssignableToTypeOf(parachaintypes.GroupIndex(0)),
+					gomock.AssignableToTypeOf(parachaintypes.SignedFullStatement{}),
+				).Return(new(Summary), nil)
+
+				return perRelayParentState{
+					table: mockTable,
+				}
+			},
+			signedStatementWithPVD: parachaintypes.SignedFullStatementWithPVD{
+				SignedFullStatement: parachaintypes.SignedFullStatement{
+					Payload: statementVDTValid,
+				},
+			},
+			mockOverseer: func(*testing.T, chan any) {},
+		},
+		{
+			description: "seconded_statement_known_candidate",
+			rpState: func() perRelayParentState {
+				ctrl := gomock.NewController(t)
+				mockTable := NewMockTable(ctrl)
+
+				mockTable.EXPECT().importStatement(
+					gomock.AssignableToTypeOf(new(tableContext)),
+					gomock.AssignableToTypeOf(parachaintypes.GroupIndex(0)),
+					gomock.AssignableToTypeOf(parachaintypes.SignedFullStatement{}),
+				).Return(new(Summary), nil)
+
+				return perRelayParentState{
+					table: mockTable,
+				}
+			},
+			perCandidate: map[parachaintypes.CandidateHash]*perCandidateState{
+				candidateHash: {
+					persistedValidationData: parachaintypes.PersistedValidationData{
+						ParentHead: parachaintypes.HeadData{
+							Data: []byte{1, 2, 3},
+						},
+					},
+					secondedLocally: false,
+					relayParent:     getDummyHash(t, 5),
+				},
+			},
+			signedStatementWithPVD: secondedSignedFullStatementWithPVD(t, statementVDTSeconded),
+			mockOverseer:           func(*testing.T, chan any) {},
+		},
+		{
+			description: "seconded_statement_unknown_candidate",
+			rpState: func() perRelayParentState {
+				ctrl := gomock.NewController(t)
+				mockTable := NewMockTable(ctrl)
+
+				mockTable.EXPECT().importStatement(
+					gomock.AssignableToTypeOf(new(tableContext)),
+					gomock.AssignableToTypeOf(parachaintypes.GroupIndex(0)),
+					gomock.AssignableToTypeOf(parachaintypes.SignedFullStatement{}),
+				).Return(new(Summary), nil)
+
+				return perRelayParentState{
+					table: mockTable,
+				}
+			},
+			perCandidate:           map[parachaintypes.CandidateHash]*perCandidateState{},
+			signedStatementWithPVD: secondedSignedFullStatementWithPVD(t, statementVDTSeconded),
+			mockOverseer: func(t *testing.T, subSystemToOverseer chan any) {
+				v := <-subSystemToOverseer
+				introduce, ok := v.(parachaintypes.ProspectiveParachainsMessageIntroduceCandidate)
+				require.True(t, ok)
+
+				introduce.Ch <- nil
+			},
+		},
+	}
+
+	for _, c := range testCases {
+		c := c
+		t.Run(c.description, func(t *testing.T) {
+			t.Parallel()
+
+			subSystemToOverseer := make(chan any)
+			defer close(subSystemToOverseer)
+
+			go c.mockOverseer(t, subSystemToOverseer)
+
+			rpState := c.rpState()
+			summary, err := rpState.importStatement(subSystemToOverseer, c.signedStatementWithPVD, c.perCandidate)
+			require.Equal(t, new(Summary), summary)
+			require.NoError(t, err)
+		})
+	}
+}
+
 func mustHexTo32BArray(t *testing.T, inputHex string) (outputArray [sr25519.PublicKeyLength]byte) {
 	t.Helper()
 	copy(outputArray[:], common.MustHexToBytes(inputHex))
@@ -158,9 +279,9 @@ func dummyTableContext(t *testing.T) tableContext {
 			index: 1,
 		},
 		groups: map[parachaintypes.CoreIndex][]parachaintypes.ValidatorIndex{
-			parachaintypes.CoreIndex{Index: 1}: {1, 2, 3},
-			parachaintypes.CoreIndex{Index: 2}: {4, 5, 6},
-			parachaintypes.CoreIndex{Index: 3}: {7, 8, 9},
+			{Index: 1}: {1, 2, 3},
+			{Index: 2}: {4, 5, 6},
+			{Index: 3}: {7, 8, 9},
 		},
 		validators: []parachaintypes.ValidatorID{
 			mustHexTo32BArray(t, "0xa262f83b46310770ae8d092147176b8b25e8855bcfbbe701d346b10db0c5385d"),
@@ -171,7 +292,7 @@ func dummyTableContext(t *testing.T) tableContext {
 	}
 }
 
-func rpStateWhenPpmDisabled(t *testing.T) perRelayParentState {
+func rpStateForSuccessfulPostImportStatement(t *testing.T, ctrl *gomock.Controller) perRelayParentState {
 	t.Helper()
 
 	attestedToReturn := attestedCandidate{
@@ -193,7 +314,6 @@ func rpStateWhenPpmDisabled(t *testing.T) perRelayParentState {
 		},
 	}
 
-	ctrl := gomock.NewController(t)
 	mockTable := NewMockTable(ctrl)
 
 	mockTable.EXPECT().drainMisbehaviors().
@@ -212,35 +332,33 @@ func rpStateWhenPpmDisabled(t *testing.T) perRelayParentState {
 }
 
 func TestPostImportStatement(t *testing.T) {
-	t.Parallel()
-
 	testCases := []struct {
 		description string
-		rpState     func() perRelayParentState
+		rpState     func(*gomock.Controller) perRelayParentState
 		summary     *Summary
+		validate    func(t *testing.T, subSystemToOverseer chan any)
 	}{
 		{
-			description: "summary_is_nil",
-			rpState: func() perRelayParentState {
-				ctrl := gomock.NewController(t)
+			description: "nil_summary",
+			rpState: func(ctrl *gomock.Controller) perRelayParentState {
 				mockTable := NewMockTable(ctrl)
-
-				mockTable.EXPECT().drainMisbehaviors().Return(map[parachaintypes.ValidatorIndex][]parachaintypes.Misbehaviour{
-					1: {parachaintypes.MultipleCandidates{}},
-				})
+				mockTable.EXPECT().drainMisbehaviors().Return(map[parachaintypes.ValidatorIndex][]parachaintypes.Misbehaviour{})
 
 				return perRelayParentState{
-					table: mockTable,
+					relayParent: common.Hash{0x01},
+					table:       mockTable,
+					backed:      make(map[parachaintypes.CandidateHash]bool),
 				}
 			},
 			summary: nil,
+			validate: func(t *testing.T, subSystemToOverseer chan any) {
+				require.Empty(t, subSystemToOverseer)
+			},
 		},
 		{
-			description: "failed_to_get_attested_candidate_from_table",
-			rpState: func() perRelayParentState {
-				ctrl := gomock.NewController(t)
+			description: "candidate_not_attested",
+			rpState: func(ctrl *gomock.Controller) perRelayParentState {
 				mockTable := NewMockTable(ctrl)
-
 				mockTable.EXPECT().drainMisbehaviors().
 					Return(map[parachaintypes.ValidatorIndex][]parachaintypes.Misbehaviour{})
 				mockTable.EXPECT().attestedCandidate(
@@ -250,23 +368,25 @@ func TestPostImportStatement(t *testing.T) {
 				).Return(nil, errors.New("could not get attested candidate from table"))
 
 				return perRelayParentState{
-					table: mockTable,
+					relayParent: common.Hash{0x01},
+					table:       mockTable,
+					backed:      make(map[parachaintypes.CandidateHash]bool),
 				}
 			},
-			summary: dummySummary(t),
+			summary: &Summary{},
+			validate: func(t *testing.T, subSystemToOverseer chan any) {
+				require.Empty(t, subSystemToOverseer)
+			},
 		},
 		{
-			description: "candidate_is_already_backed",
-			rpState: func() perRelayParentState {
-				ctrl := gomock.NewController(t)
-				mockTable := NewMockTable(ctrl)
+			description: "candidate_already_backed",
+			rpState: func(ctrl *gomock.Controller) perRelayParentState {
 
 				candidate := getDummyCommittedCandidateReceipt(t)
-				hash, err := candidate.Hash()
+				candidateHash, err := parachaintypes.GetCandidateHash(candidate)
 				require.NoError(t, err)
 
-				candidateHash := parachaintypes.CandidateHash{Value: hash}
-
+				mockTable := NewMockTable(ctrl)
 				mockTable.EXPECT().drainMisbehaviors().
 					Return(map[parachaintypes.ValidatorIndex][]parachaintypes.Misbehaviour{})
 				mockTable.EXPECT().attestedCandidate(
@@ -277,70 +397,48 @@ func TestPostImportStatement(t *testing.T) {
 					groupID:                   4,
 					committedCandidateReceipt: candidate,
 				}, nil)
-
 				return perRelayParentState{
-					table: mockTable,
+					relayParent: common.Hash{0x01},
+					table:       mockTable,
 					backed: map[parachaintypes.CandidateHash]bool{
 						candidateHash: true,
 					},
 				}
 			},
-			summary: dummySummary(t),
+			summary: &Summary{},
+			validate: func(t *testing.T, subSystemToOverseer chan any) {
+				require.Empty(t, subSystemToOverseer)
+			},
 		},
 		{
-			description: "Validity_vote_from_unknown_validator",
-			rpState: func() perRelayParentState {
-				ctrl := gomock.NewController(t)
-				mockTable := NewMockTable(ctrl)
-
-				mockTable.EXPECT().drainMisbehaviors().
-					Return(map[parachaintypes.ValidatorIndex][]parachaintypes.Misbehaviour{})
-				mockTable.EXPECT().attestedCandidate(
-					gomock.AssignableToTypeOf(parachaintypes.CandidateHash{}),
-					gomock.AssignableToTypeOf(new(tableContext)),
-					gomock.AssignableToTypeOf(uint32(0)),
-				).Return(&attestedCandidate{
-					groupID:                   3,
-					committedCandidateReceipt: getDummyCommittedCandidateReceipt(t),
-				}, nil)
-
-				return perRelayParentState{
-					table:        mockTable,
-					backed:       map[parachaintypes.CandidateHash]bool{},
-					tableContext: dummyTableContext(t),
-				}
+			description: "successful_post_import_statement",
+			rpState: func(ctrl *gomock.Controller) perRelayParentState {
+				return rpStateForSuccessfulPostImportStatement(t, ctrl)
 			},
-			summary: dummySummary(t),
-		},
-		{
-			description: "prospective_parachain_mode_is_disabled",
-			rpState: func() perRelayParentState {
-				return rpStateWhenPpmDisabled(t)
+			summary: &Summary{},
+			validate: func(t *testing.T, subSystemToOverseer chan any) {
+				require.Len(t, subSystemToOverseer, 3)
+				require.IsType(t, parachaintypes.ProspectiveParachainsMessageCandidateBacked{}, <-subSystemToOverseer)
+				require.IsType(t, collatorprotocolmessages.Backed{}, <-subSystemToOverseer)
+				require.IsType(t, statementdistributionmessages.Backed{}, <-subSystemToOverseer)
 			},
-			summary: dummySummary(t),
-		},
-		{
-			description: "prospective_parachain_mode_is_enabled",
-			rpState: func() perRelayParentState {
-				state := rpStateWhenPpmDisabled(t)
-				return state
-			},
-			summary: dummySummary(t),
 		},
 	}
 
-	for _, c := range testCases {
-		c := c
-		t.Run(c.description, func(t *testing.T) {
+	for _, tc := range testCases {
+		tc := tc
+		t.Run(tc.description, func(t *testing.T) {
 			t.Parallel()
 
-			subSystemToOverseer := make(chan any)
+			ctrl := gomock.NewController(t)
+			defer ctrl.Finish()
+
+			subSystemToOverseer := make(chan any, 10)
 			defer close(subSystemToOverseer)
 
-			go mockOverseer(t, subSystemToOverseer)
-
-			rpState := c.rpState()
-			rpState.postImportStatement(subSystemToOverseer, c.summary)
+			rpState := tc.rpState(ctrl)
+			rpState.postImportStatement(subSystemToOverseer, tc.summary)
+			tc.validate(t, subSystemToOverseer)
 		})
 	}
 }
@@ -660,10 +758,8 @@ func TestHandleStatementMessage(t *testing.T) {
 	err := statementVDTSeconded.SetValue(seconded)
 	require.NoError(t, err)
 
-	hash, err := dummyCCR.Hash()
+	candidateHash, err := parachaintypes.GetCandidateHash(dummyCCR)
 	require.NoError(t, err)
-
-	candidateHash := parachaintypes.CandidateHash{Value: hash}
 
 	statementVDTValid := parachaintypes.NewStatementVDT()
 	err = statementVDTValid.SetValue(parachaintypes.Valid(candidateHash))
@@ -676,7 +772,6 @@ func TestHandleStatementMessage(t *testing.T) {
 		signedStatementWithPVD parachaintypes.SignedFullStatementWithPVD
 		err                    string
 	}{
-
 		{
 			description: "unknown_relay_parent",
 			perRelayParent: func() map[common.Hash]*perRelayParentState {
@@ -732,7 +827,6 @@ func TestHandleStatementMessage(t *testing.T) {
 			},
 			err: "",
 		},
-
 		{
 			description: "paraId_is_not_assigned_to_the_local_validator",
 			perRelayParent: func() map[common.Hash]*perRelayParentState {
@@ -756,7 +850,12 @@ func TestHandleStatementMessage(t *testing.T) {
 
 				return map[common.Hash]*perRelayParentState{
 					relayParent: {
-						table: mockTable,
+						table:        mockTable,
+						assignedCore: &parachaintypes.CoreIndex{Index: 1},
+						fallbacks: map[parachaintypes.CandidateHash]attestingData{
+							candidateHash: {},
+						},
+						tableContext: dummyTableContext(t),
 					},
 				}
 			},
@@ -767,7 +866,6 @@ func TestHandleStatementMessage(t *testing.T) {
 			},
 			err: "",
 		},
-
 		{
 			description: "statementVDT_set_to_valid_and_candidate_not_in_fallbacks",
 			perRelayParent: func() map[common.Hash]*perRelayParentState {
@@ -792,9 +890,10 @@ func TestHandleStatementMessage(t *testing.T) {
 
 				return map[common.Hash]*perRelayParentState{
 					relayParent: {
-						table:     mockTable,
-						backed:    map[parachaintypes.CandidateHash]bool{},
-						fallbacks: map[parachaintypes.CandidateHash]attestingData{},
+						assignedCore: &parachaintypes.CoreIndex{Index: 4},
+						table:        mockTable,
+						backed:       map[parachaintypes.CandidateHash]bool{},
+						fallbacks:    map[parachaintypes.CandidateHash]attestingData{},
 					},
 				}
 			},
@@ -805,7 +904,6 @@ func TestHandleStatementMessage(t *testing.T) {
 			},
 			err: errFallbackNotAvailable.Error(),
 		},
-
 		{
 			description: "statementVDT_set_to_valid_also_same_validatorIndex_in_tableContext_and_signedStatement",
 			perRelayParent: func() map[common.Hash]*perRelayParentState {
@@ -830,6 +928,7 @@ func TestHandleStatementMessage(t *testing.T) {
 
 				return map[common.Hash]*perRelayParentState{
 					relayParent: {
+						assignedCore: &parachaintypes.CoreIndex{Index: 4},
 						table:        mockTable,
 						tableContext: dummyTableContext(t),
 						backed:       map[parachaintypes.CandidateHash]bool{},
@@ -871,6 +970,7 @@ func TestHandleStatementMessage(t *testing.T) {
 
 				return map[common.Hash]*perRelayParentState{
 					relayParent: {
+						assignedCore: &parachaintypes.CoreIndex{Index: 4},
 						table:        mockTable,
 						tableContext: dummyTableContext(t),
 						backed:       map[parachaintypes.CandidateHash]bool{},
@@ -915,6 +1015,7 @@ func TestHandleStatementMessage(t *testing.T) {
 
 				return map[common.Hash]*perRelayParentState{
 					relayParent: {
+						assignedCore: &parachaintypes.CoreIndex{Index: 4},
 						table:        mockTable,
 						tableContext: dummyTableContext(t),
 						backed:       map[parachaintypes.CandidateHash]bool{},
@@ -972,7 +1073,8 @@ func TestHandleStatementMessage(t *testing.T) {
 
 				return map[common.Hash]*perRelayParentState{
 					relayParent: {
-						table: mockTable,
+						assignedCore: &parachaintypes.CoreIndex{Index: 4},
+						table:        mockTable,
 						backed: map[parachaintypes.CandidateHash]bool{
 							candidateHash: true,
 						},
@@ -1013,7 +1115,8 @@ func TestHandleStatementMessage(t *testing.T) {
 
 				return map[common.Hash]*perRelayParentState{
 					relayParent: {
-						table: mockTable,
+						assignedCore: &parachaintypes.CoreIndex{Index: 4},
+						table:        mockTable,
 						backed: map[parachaintypes.CandidateHash]bool{
 							candidateHash: true,
 						},
@@ -1029,7 +1132,6 @@ func TestHandleStatementMessage(t *testing.T) {
 			err:                    "",
 		},
 	}
-
 	for _, c := range testCases {
 		c := c
 		t.Run(c.description, func(t *testing.T) {
